refactor: use runtime dir
