+++ src/index.ts
@@ -1,1 +1,2 @@
-export * from './runtime/composables'
+/* dummy entry for types field */
+export * from './runtime'
